LayoutTests/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/navigating-across-documents/initial-empty-document/iframe-src-204-fragment-expected.txt


FAIL src assert_equals: history.length must not change after normal navigation from initial empty document expected 2 but got 3
FAIL location.href assert_equals: history.length must not change after normal navigation from initial empty document expected 3 but got 4
FAIL location.assign assert_equals: history.length must not change after normal navigation from initial empty document expected 4 but got 5
FAIL window.open assert_equals: history.length must not change after normal navigation from initial empty document expected 5 but got 6
FAIL link click assert_equals: history.length must not change after fragment navigation on initial empty document expected 6 but got 7
